+++ src/nativeMain/kotlin/model/Author.kt
@@ -1,0 +1,10 @@
+package model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Author(
+    val tag: String,
+    val name: String,
+    val email: String
+)
