+++ src/nativeMain/kotlin/model/SignatureFormatter.kt
@@ -1,0 +1,5 @@
+package model
+
+interface SignatureFormatter {
+    fun generateSignatures(authors: List<Author>): String
+}
